requirements.txt
addict==2.4.0
appdirs==1.4.4
apturl==0.5.2
attrs==21.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
contourpy==1.0.7
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
Cython==0.29.33
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info===1.1build1
docker-pycreds==0.4.0
evdev==1.6.1
exceptiongroup==1.1.0
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.30
graphviz==0.20.1
httplib2==0.20.2
idna==3.3
imagesize==1.4.1
importlib-metadata==4.6.4
iniconfig==2.0.0
jeepney==0.7.1
jsonschema==3.2.0
keyring==23.5.0
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.6.3
mmcls==0.25.0
mmcv-full==1.7.0
-e git+https://github.com/open-mmlab/mmdetection.git@c14dd6c42efb63f662a63fe403198bac82f47aa6#egg=mmdet
-e git+https://github.com/open-mmlab/mmsegmentation.git@8ea3ea39ee747f3393b3798deb62b1ff36f8bcde#egg=mmsegmentation
more-itertools==8.10.0
netifaces==0.11.0
nftables==0.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.0
olefile==0.46
opencv-python==4.7.0.68
packaging==23.0
pandas==1.5.3
pathtools==0.1.2
pexpect==4.8.0
Pillow==9.0.1
pluggy==1.0.0
prettytable==3.6.0
protobuf==3.12.4
psutil==5.9.4
ptyprocess==0.7.0
pycairo==1.20.1
pycocotools==2.0.6
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pynput==1.7.6
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
pytest==7.2.1
python-apt==2.4.0
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-xlib==0.33
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
scipy==1.10.0
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
sentry-sdk==1.15.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
systemd-python==234
terminaltables==3.1.10
tomli==2.0.1
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
torchviz==0.0.2
typing_extensions==4.4.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
wadllib==1.3.6
wandb==0.13.10
wcwidth==0.2.6
xdg==5
xkit==0.0.0
yapf==0.32.0
zipp==1.0.0
